feat: upgrade node to v20 (#95)
The node18 runner is not going to be supported by Github, and the
current Node LTS version is 18. So let's jump directly from v16 to v20,
to keep up with Node LTS.

BREAKING CHANGE: The action now runs on node20 runner. If you use
self-hosted runners, make sure you are able to run actions with node
v20.